Bug corrections
As i was using, i passed insertFromTop = false, and it keeps puting on problem is 'insertFromTop ? insertFromTop : true' alwas goes to TRUE since, false arguments will fall on false stament on ternary operator witch was true;

and the other corrections is, since scope is created manualy sometime angular does not update views, so it needs to be updated manualy. and then show all toasts, it just happens if toasts are beeing spammed, but as i'm using it with socket notifications it would be a problem.